Add files via upload
This is the initial solution concept developed in python.  Using backward induction, start with all the possible end states of the game where Player 2 loses. Then produce the set of Winning game states in the previous round.  Next produce the Losing states the the round prior, where for each hand, all possible moves result in a Winning game state. And continue.